requirements.txt
flake8
Appium-Python-Client
pytest
